Changed the cache back to slices because maps are sorted by keys. Can now reorder lists and items

+++ src/internal/storage/storage.go
@@ -2,6 +2,7 @@ package storage
 
 import (
 	"errors"
+	"slices"
 
 	"github.com/BeringLogic/flimsy/internal/db"
 )
@@ -10,14 +11,14 @@ import (
 type FlimsyStorage struct {
   db *db.FlimsyDB
   Config *db.Config
-  Lists map[int64]*db.List
-  Items map[int64]*db.Item
-  ListsAndItems map[int64]*listAndItems
+  Lists []*db.List
+  Items []*db.Item
+  AllListsAndItems []*listAndItems
 }
 
 type listAndItems struct {
   List *db.List
-  Items map[int64]*db.Item
+  Items []*db.Item
 }
 
 
@@ -42,7 +43,7 @@ func (storage *FlimsyStorage) Init() error {
   }
 
   storage.Config, err = storage.db.LoadConfig(); if err != nil {
-  if err = storage.db.Seed(); err != nil {
+    if err = storage.db.Seed(); err != nil {
       return errors.Join(SeedError, err);
     }
     if storage.Config, err = storage.db.LoadConfig(); err != nil {
@@ -58,26 +59,26 @@ func (storage *FlimsyStorage) Init() error {
       return errors.Join(LoadItemsError, err);
   }
 
-  storage.ListsAndItems = storage.getAllListsAndItems()
+  storage.AllListsAndItems = storage.getAllListsAndItems()
 
   return nil
 }
 
-func (flimsyStorage *FlimsyStorage) getAllListsAndItems() map[int64]*listAndItems {
-  AllListsAndItems := make(map[int64]*listAndItems, 0)
+func (flimsyStorage *FlimsyStorage) getAllListsAndItems() []*listAndItems {
+  AllListsAndItems := make([]*listAndItems, 0)
   for _, list := range flimsyStorage.Lists {
     lai := listAndItems {
       List: list,
-      Items: make(map[int64]*db.Item),
+      Items: make([]*db.Item, 0),
     }
 
     for _, item := range flimsyStorage.Items {
       if item.List_id == list.Id {
-        lai.Items[item.Id] = item
+        lai.Items = append(lai.Items, item)
       }
     }
 
-    AllListsAndItems[list.Id] = &lai
+    AllListsAndItems = append(AllListsAndItems, &lai)
   }
   
   return AllListsAndItems
@@ -96,13 +97,25 @@ func (flimsyStorage *FlimsyStorage) AddList(title string, number_of_rows int) (*
     return nil, err
   }
 
-  flimsyStorage.Lists[list.Id] = list
-  flimsyStorage.ListsAndItems[list.Id] = &listAndItems {
+  flimsyStorage.Lists = append(flimsyStorage.Lists, list)
+
+  lai := listAndItems {
     List: list,
-    Items: make(map[int64]*db.Item),
+    Items: make([]*db.Item, 0),
+  }
+  flimsyStorage.AllListsAndItems = append(flimsyStorage.AllListsAndItems, &lai)
+
+  return &lai, nil
+}
+
+func (flimsyStorage *FlimsyStorage) GetList(id int64) (*db.List, error) {
+  for _, list := range flimsyStorage.Lists {
+    if list.Id == id {
+      return list, nil
+    }
   }
 
-  return flimsyStorage.ListsAndItems[list.Id], nil
+  return nil, errors.New("List not found")
 }
 
 func (flimsyStorage *FlimsyStorage) SaveList(list *db.List) (*listAndItems, error) {
@@ -110,22 +123,53 @@ func (flimsyStorage *FlimsyStorage) SaveList(list *db.List) (*listAndItems, erro
     return nil, err
   }
 
-  return flimsyStorage.ListsAndItems[list.Id], nil
-}
-
-func (flimsyStorage *FlimsyStorage) DeleteList(id int64) error {
-  for _, item := range flimsyStorage.ListsAndItems[id].Items {
-    if err := flimsyStorage.DeleteItem(item.Id); err != nil {
-      return err
+  var lai *listAndItems
+  for _, n := range flimsyStorage.AllListsAndItems {
+    if n.List.Id == list.Id {
+      lai = n
     }
   }
 
-  if err := flimsyStorage.db.DeleteList(id); err != nil {
+  return lai, nil
+}
+
+func (flimsyStorage *FlimsyStorage) DeleteList(id int64) error {
+  for i, n := range flimsyStorage.AllListsAndItems {
+    if n.List.Id == id {
+      for _, item := range n.Items {
+        if err := flimsyStorage.DeleteItem(item.Id); err != nil {
+          return err
+        }
+      }
+      if err := flimsyStorage.db.DeleteList(id); err != nil {
+        return err
+      }
+
+      flimsyStorage.AllListsAndItems = slices.Delete(flimsyStorage.AllListsAndItems, i, i+1)
+      flimsyStorage.Lists = slices.Delete(flimsyStorage.Lists, i, i+1)
+      break
+    }
+  }
+
+  return nil
+}
+
+func (flimsyStorage *FlimsyStorage) ReorderLists(list_ids []int64) error {
+  if err := flimsyStorage.db.ReorderLists(list_ids); err != nil {
     return err
   }
 
-  delete(flimsyStorage.ListsAndItems, id)
-  delete(flimsyStorage.Lists, id)
+  var loadListErr error
+  if flimsyStorage.Lists, loadListErr = flimsyStorage.db.LoadLists(); loadListErr != nil {
+    return errors.Join(LoadListsError, loadListErr);
+  }
+
+  var loadItemsErr error
+  if flimsyStorage.Items, loadItemsErr = flimsyStorage.db.LoadItems(); loadItemsErr  != nil {
+    return errors.Join(LoadItemsError, loadItemsErr);
+  }
+
+  flimsyStorage.AllListsAndItems = flimsyStorage.getAllListsAndItems()
 
   return nil
 }
@@ -135,10 +179,25 @@ func (flimsyStorage *FlimsyStorage) AddItem(list_id int64, title string, url str
     return nil, err
   }
 
-  flimsyStorage.Items[item.Id] = item
-  flimsyStorage.ListsAndItems[list_id].Items[item.Id] = item
+  flimsyStorage.Items = append(flimsyStorage.Items, item)
+
+  for _, lislistAndItems := range flimsyStorage.AllListsAndItems {
+    if lislistAndItems.List.Id == list_id {
+      lislistAndItems.Items = append(lislistAndItems.Items, item)
+    }
+  }
 
   return item, nil
+}
+
+func (flimsyStorage *FlimsyStorage) GetItem(id int64) (*db.Item, error) {
+  for _, item := range flimsyStorage.Items {
+    if item.Id == id {
+      return item, nil
+    }
+  }
+
+  return nil, errors.New("Item not found")
 }
 
 func (flimsyStorage *FlimsyStorage) SaveItem(item *db.Item) error {
@@ -150,13 +209,28 @@ func (flimsyStorage *FlimsyStorage) DeleteItem(id int64) error {
     return err
   }
 
-  for _, listAndItems := range flimsyStorage.ListsAndItems {
-    if listAndItems.Items[id] != nil {
-      delete(listAndItems.Items, id)
+  for _, listAndItems := range flimsyStorage.AllListsAndItems {
+    for i, item := range listAndItems.Items {
+      if item.Id == id {
+        listAndItems.Items = slices.Delete(listAndItems.Items, i, i+1)
+        return nil
+      }
     }
   }
 
-  delete(flimsyStorage.Items, id)
+  return errors.New("Item not found in cache")
+}
+
+func (flimsyStorage *FlimsyStorage) ReorderItems(list_id int64, item_ids []int64) error {
+  err := flimsyStorage.db.ReorderItems(list_id, item_ids); if err != nil {
+    return err
+  }
+
+  if flimsyStorage.Items, err = flimsyStorage.db.LoadItems(); err != nil {
+    return errors.Join(LoadItemsError, err);
+  }
+
+  flimsyStorage.AllListsAndItems = flimsyStorage.getAllListsAndItems()
 
   return nil
 }
